Utilities\BitField.h: Try to fix some warnings
Cleanup and make bitfield extraction u32 if possible

+++ rpcs3/Emu/Memory/vm.cpp
@@ -570,7 +570,7 @@ namespace vm
 						}
 
 						addr2 += size3;
-						size2 -= size3;
+						size2 -= static_cast<u32>(size3);
 					}
 
 					return 0;
